+++ .github/files/lint-project-structure.sh
@@ -407,6 +407,31 @@ for PROJECT in projects/*/*; do
 		fi
 	fi
 
+	# - `.extra.dependencies.test-only` must refer to dev dependencies.
+	if jq -e '.extra.dependencies["test-only"] // empty' "$PROJECT/composer.json" >/dev/null; then
+		while IFS=$'\t' read -r LINE DEP; do
+			if [[ ! -e "projects/$DEP/composer.json" ]]; then
+				EXIT=1
+				echo "::error file=$PROJECT/composer.json,line=${LINE}::Dependency \"$DEP\" does not exist in the monorepo."
+			elif [[ "$DEP" == packages/* ]]; then
+				N=$( jq -r .name "projects/$DEP/composer.json" )
+				if ! jq -e --arg N "$N" '.["require-dev"][$N]' "$PROJECT/composer.json" &>/dev/null; then
+					EXIT=1
+					echo "::error file=$PROJECT/composer.json,line=${LINE}::Project \"$DEP\" ($N) is not a dev dependency of $SLUG."
+				fi
+			elif [[ "$DEP" == js-packages/* ]]; then
+				N=$( jq -r .name "projects/$DEP/package.json" 2>/dev/null || true )
+				if ! jq -e --arg N "$N" '.devDependencies[$N]' "$PROJECT/package.json" &>/dev/null; then
+					EXIT=1
+					echo "::error file=$PROJECT/composer.json,line=${LINE}::Project \"$DEP\" ($N) is not a dev dependency of $SLUG."
+				fi
+			else
+				EXIT=1
+				echo "::error file=$PROJECT/composer.json,line=${LINE}::Dependency \"$DEP\" is neither a package nor a js-package."
+			fi
+		done < <( jq --stream -r 'if length == 2 and .[0][:-1] == ["extra","dependencies","test-only"] then [input_line_number,.[1]] | @tsv else empty end' "$PROJECT/composer.json" )
+	fi
+
 done
 
 # - Monorepo root composer.json must also use dev deps appropriately.