+++ projects/packages/explat/src/class-explat.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * Package description here
+ *
+ * @package automattic/jetpack-explat
+ */
+
+namespace Automattic\Jetpack;
+
+/**
+ * Class description.
+ */
+class Explat {
+
+	const PACKAGE_VERSION = '0.1.0-alpha';
+}