+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/ai-image/index.ts
@@ -1,4 +1,5 @@
-import FeaturedImage, { FEATURED_IMAGE_PLACEMENT_MEDIA_SOURCE_DROPDOWN } from './featured-image';
+import FeaturedImage from './featured-image';
 import GeneralPurposeImage from './general-purpose-image';
+import { PLACEMENT_MEDIA_SOURCE_DROPDOWN } from './types';
 
-export { FeaturedImage, FEATURED_IMAGE_PLACEMENT_MEDIA_SOURCE_DROPDOWN, GeneralPurposeImage };
+export { FeaturedImage, PLACEMENT_MEDIA_SOURCE_DROPDOWN, GeneralPurposeImage };