+++ tools/cli/commands/generate.js
@@ -958,7 +958,7 @@ function createReadMeTxt( answers ) {
 		'Tags: jetpack, stuff\n' +
 		'Requires at least: 6.4\n' +
 		'Requires PHP: 7.0\n' +
-		'Tested up to: 6.5\n' +
+		'Tested up to: 6.6\n' +
 		`Stable tag: ${ answers.version }\n` +
 		'License: GPLv2 or later\n' +
 		'License URI: http://www.gnu.org/licenses/gpl-2.0.html\n' +