+++ tools/e2e-commons/pages/frontend/postFrontend.js
@@ -1,5 +1,5 @@
-import WpPage from '../wp-page.js';
 import { resolveSiteUrl } from '../../helpers/utils-helper.js';
+import WpPage from '../wp-page.js';
 
 export default class PostFrontendPage extends WpPage {
 	constructor( page ) {
@@ -10,8 +10,9 @@ export default class PostFrontendPage extends WpPage {
 	/**
 	 * Checks whether specific block is rendered on frontend. All the custom logic is defined in block's `isRendered` static method
 	 *
-	 * @param {Object} BlockClass Block class that has a static `isRendered` method
-	 * @param {Object} args       An object of any additional instance values required by the class `isRendered` method
+	 * @param {object} BlockClass - Block class that has a static `isRendered` method
+	 * @param {object} args       - An object of any additional instance values required by the class `isRendered` method
+	 * @returns {boolean} Always true.
 	 */
 	async isRenderedBlockPresent( BlockClass, args ) {
 		await BlockClass.isRendered( this.page, args );