+++ projects/packages/masterbar/CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2024-06-26
+### Changed
+- Internal updates.
+
+## [0.2.3] - 2024-06-26
+### Fixed
+- For sites with Classic view, don't load the masterbar package except the admin color schemes functionality. [#38020]
+
+## [0.2.2] - 2024-06-25
+### Fixed
+- Masterbar: Fix missing private badge on admin menu for private WoA sites [#38026]
+
+## [0.2.1] - 2024-06-25
+### Fixed
+- Profile: Restore profile fields on Classic interface [#38016]
+
+## [0.2.0] - 2024-06-21
+### Changed
+- Masterbar: Require and use 'jetpack-masterbar' package in jetpack-mu-wpcom [#37812]
+
+### Fixed
+- Masterbar: Remove old fly panel [#37764]
+
 ## [0.1.1] - 2024-06-17
 ### Changed
 - Updated package dependencies. [#37796]
@@ -23,4 +46,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated package dependencies. [#37669]
 - Updated package dependencies. [#37706]
 
+[0.2.4]: https://github.com/Automattic/jetpack-masterbar/compare/v0.2.3...v0.2.4
+[0.2.3]: https://github.com/Automattic/jetpack-masterbar/compare/v0.2.2...v0.2.3
+[0.2.2]: https://github.com/Automattic/jetpack-masterbar/compare/v0.2.1...v0.2.2
+[0.2.1]: https://github.com/Automattic/jetpack-masterbar/compare/v0.2.0...v0.2.1
+[0.2.0]: https://github.com/Automattic/jetpack-masterbar/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/Automattic/jetpack-masterbar/compare/v0.1.0...v0.1.1