+++ projects/packages/sync/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.3] - 2024-06-25
+### Changed
+- Internal updates.
+
+## [3.1.2] - 2024-06-21
+### Added
+- Adding a new toggles to control the newsletter byline [#37916]
+
 ## [3.1.1] - 2024-06-13
 ### Changed
 - Internal updates.
@@ -1180,6 +1188,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Packages: Move sync to a classmapped package
 
+[3.1.3]: https://github.com/Automattic/jetpack-sync/compare/v3.1.2...v3.1.3
+[3.1.2]: https://github.com/Automattic/jetpack-sync/compare/v3.1.1...v3.1.2
 [3.1.1]: https://github.com/Automattic/jetpack-sync/compare/v3.1.0...v3.1.1
 [3.1.0]: https://github.com/Automattic/jetpack-sync/compare/v3.0.2...v3.1.0
 [3.0.2]: https://github.com/Automattic/jetpack-sync/compare/v3.0.1...v3.0.2